+++ e3nn_jax/__init__.py
@@ -73,7 +73,7 @@ from e3nn_jax._src.reduced_tensor_product import (
     reduced_symmetric_tensor_product_basis,
     reduced_antisymmetric_tensor_product_basis,
 )
-from e3nn_jax._src.s2grid import s2_irreps, to_s2grid, to_s2point, from_s2grid, s2_sum_of_diracs, SphericalSignal
+from e3nn_jax._src.s2grid import s2_irreps, to_s2grid, to_s2point, from_s2grid, s2_dirac, SphericalSignal
 
 # make submodules flax and haiku available
 from e3nn_jax import flax, haiku
@@ -168,7 +168,7 @@ __all__ = [
     "to_s2grid",
     "to_s2point",
     "from_s2grid",
-    "s2_sum_of_diracs",
+    "s2_dirac",
     "SphericalSignal",
     "flax",
     "haiku",